\name{coeff,bslda-method}
\alias{coeff,bslda-method}
\alias{coeff.bslda}
\docType{methods}

\title{bslda result}
\usage{
\S4method{coeff}{bslda}(object)
}
\arguments{
\item{object}{Output of bslda function}
}
\description{
Provide coeff coefficients, accuracy, and balanced accuracy.
}
