\name{predict,bslda-method}
\alias{predict,bslda-method}
\alias{predict.bslda}
\docType{methods}

\title{bslda result}
\usage{
\S4method{predict}{bslda}(object)
}
\arguments{
\item{object}{Output of bslda function}
\item{x}{Data matrix of new individuals}
\item{y}{Class vector of new individuals}
}
\description{
Provide predicted class of new individuals.
}
